+++ index.ts
@@ -10,8 +10,6 @@ export * from './src/utils/getChapterRangeWordCount';
 export * from './src/utils/getSubsetFromBookRange';
 export * from './src/utils/isBookJoinable';
 export * from './src/utils/isBookPartOfGroup';
-export * from './src/utils/normaliseChapter';
 export * from './src/utils/normaliseChapterRange';
-export * from './src/utils/normaliseVerse';
 
 export { Bible as default } from './src/Bible';