+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 
 -  Added Passage class
 -  Started changelog
+-  Deprecated `normaliseChapter` util, which is now accessed from `Passage.normaliseChapter`
+-  Deprecated `normaliseVerse` util, which is now accessed from `Passage.normaliseVerse`